refactor: 클래스명을 보다 명확하게 변경, 패키지 구조 개선
