+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/auth/factory/userinfo/OAuth2UserInfo.java
@@ -1,0 +1,9 @@
+package com.mars.NangPaGo.auth.factory.userinfo;
+
+public interface OAuth2UserInfo {
+    String getProvider();
+    String getProviderId();
+    String getEmail();
+    String getName();
+    String getProfileImageUrl();
+}
