+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/common/aop/auth/AuthenticatedUser.java
@@ -1,0 +1,12 @@
+package com.mars.NangPaGo.common.aop.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface AuthenticatedUser {
+
+}
