+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/common/component/auth/AuthenticationHolder.java
@@ -1,0 +1,19 @@
+package com.mars.NangPaGo.common.component.auth;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthenticationHolder {
+    private AuthenticationHolder() { }
+
+    public static String getCurrentUserEmail() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null && authentication.getPrincipal() instanceof UserDetails userDetails) {
+            return userDetails.getUsername();
+        }
+        return "anonymous_user";
+    }
+}
