+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/common/util/JwtUtil.java
@@ -1,0 +1,121 @@
+package com.mars.NangPaGo.common.util;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.Jwts.SIG;
+import javax.crypto.spec.SecretKeySpec;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.Authentication;
+import org.springframework.stereotype.Component;
+
+import javax.crypto.SecretKey;
+import java.nio.charset.StandardCharsets;
+import java.util.Date;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.stream.Collectors;
+import org.springframework.security.core.GrantedAuthority;
+
+@Component
+public class JwtUtil {
+
+    private final SecretKey secretKey;
+    private final long accessTokenExpireMillis;
+    private final long refreshTokenExpireMillis;
+
+    public JwtUtil(
+        @Value("${jwt.secret}") String secret,
+        @Value("${jwt.token.access-expiration-time}") Long accessTokenExpireMillis,
+        @Value("${jwt.token.refresh-expiration-time}") Long refreshTokenExpireMillis
+    ) {
+        this.secretKey = new SecretKeySpec(secret.getBytes(StandardCharsets.UTF_8), SIG.HS256.key().build().getAlgorithm());
+        this.accessTokenExpireMillis = accessTokenExpireMillis;
+        this.refreshTokenExpireMillis = refreshTokenExpireMillis;
+    }
+
+    public String getCategory(String token) {
+        return Jwts.parser()
+            .verifyWith(secretKey)
+            .build()
+            .parseSignedClaims(token)
+            .getPayload()
+            .get("category", String.class);
+    }
+
+    public String getEmail(String token) {
+        return Jwts.parser()
+            .verifyWith(secretKey)
+            .build()
+            .parseSignedClaims(token)
+            .getPayload()
+            .get("email", String.class);
+    }
+
+    public String getRole(String token) {
+        return Jwts.parser()
+            .verifyWith(secretKey)
+            .build()
+            .parseSignedClaims(token)
+            .getPayload()
+            .get("role", String.class);
+    }
+
+    public Boolean isExpired(String token) {
+        try {
+            Date expiration = Jwts.parser()
+                .verifyWith(secretKey)
+                .build()
+                .parseSignedClaims(token)
+                .getPayload()
+                .getExpiration();
+            return expiration.before(new Date());
+        } catch (ExpiredJwtException e) {
+            return true;
+        }
+    }
+
+    public String createJwt(String category, String email, String role, Long expiredMs) {
+        return Jwts.builder()
+            .claim("category", category)
+            .claim("email", email)
+            .claim("role", role)
+            .issuedAt(new Date(System.currentTimeMillis()))
+            .expiration(new Date(System.currentTimeMillis() + expiredMs))
+            .signWith(secretKey)
+            .compact();
+    }
+
+    public Authentication getAuthentication(String token) {
+        Claims claims = parseClaims(token);
+        
+        Collection<? extends GrantedAuthority> authorities = 
+            Arrays.stream(new String[]{claims.get("role").toString()})
+                .map(SimpleGrantedAuthority::new)
+                .collect(Collectors.toList());
+        
+        UserDetails principal = new User(claims.get("email").toString(), "", authorities);
+        
+        return new UsernamePasswordAuthenticationToken(principal, "", authorities);
+    }
+
+    private Claims parseClaims(String token) {
+        return Jwts.parser()
+            .verifyWith(secretKey)
+            .build()
+            .parseSignedClaims(token)
+            .getPayload();
+    }
+
+    public long getAccessTokenExpireMillis() {
+        return accessTokenExpireMillis;
+    }
+
+    public long getRefreshTokenExpireMillis() {
+        return refreshTokenExpireMillis;
+    }
+}
