--- NangPaGo-be/src/main/java/com/mars/NangPaGo/config/JpaAuditingConfig.java
@@ -1,10 +1,0 @@
-package com.mars.NangPaGo.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {
-
-}
