+++ NangPaGo-be/src/main/java/com/mars/NangPaGo/config/jpa/JpaAuditingConfig.java
@@ -1,0 +1,10 @@
+package com.mars.NangPaGo.config.jpa;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {
+
+}
