--- NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/auth/aop/AuthenticatedUser.java
@@ -1,12 +1,0 @@
-package com.mars.NangPaGo.domain.auth.aop;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface AuthenticatedUser {
-
-}
