--- NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/auth/factory/userinfo/OAuth2UserInfo.java
@@ -1,9 +1,0 @@
-package com.mars.NangPaGo.domain.auth.factory.userinfo;
-
-public interface OAuth2UserInfo {
-    String getProvider();
-    String getProviderId();
-    String getEmail();
-    String getName();
-    String getProfileImageUrl();
-}
