--- NangPaGo-be/src/main/java/com/mars/NangPaGo/domain/jwt/dto/RefreshTokenDto.java
@@ -1,23 +1,0 @@
-package com.mars.NangPaGo.domain.jwt.dto;
-
-import com.mars.NangPaGo.domain.jwt.entity.RefreshToken;
-
-import java.time.LocalDateTime;
-
-public record RefreshTokenDto(
-    String email,
-    String refreshToken,
-    LocalDateTime expiration
-) {
-    public static RefreshTokenDto from(RefreshToken refreshToken) {
-        return new RefreshTokenDto(
-            refreshToken.getEmail(),
-            refreshToken.getRefreshToken(),
-            refreshToken.getExpiration()
-        );
-    }
-
-    public RefreshToken toEntity() {
-        return RefreshToken.create(this.email, this.refreshToken, this.expiration);
-    }
-}
